refactor(server): extract mboxrd codec into retrieval/mbox.rs
`export_messages.rs` intermixed MCP validation, export orchestration,
and a self-contained mboxrd serializer that knows nothing about IMAP,
accounts, sandboxes, or MCP.

Move the codec (`MBOX_SEPARATOR`, `build_mbox`, `escape_from_lines_into`,
`write_mbox_line`, `line_is_from`) and its `build_mbox_tests` /
`git_am_tests` modules verbatim into a new `retrieval/mbox.rs`.
`build_mbox` is `pub(super)` — its only caller is the export handler;
the helpers and the separator stay module-private. The From-line
escaping security property is byte-for-byte unchanged.

Closes #331
