+++ crates/rimap-server/src/tools/retrieval/mod.rs
@@ -4,6 +4,7 @@ pub mod download_attachment;
 pub mod export_messages;
 pub mod fetch_message;
 pub mod list_attachments;
+mod mbox;
 pub(crate) mod part_walker;
 pub(crate) mod sandbox;
 pub mod search;
